Make the library heading the mode switcher in one pinned header
The Albums/Composers switcher was a small dropdown tucked into each mode's own
header, and those headers lived in different places (the album header scrolled
inside the grid; the composer header sat above the narrow list pane), so it
never stayed put across a mode switch.

Lift a single header above the content: the big bold heading itself is the
switcher — it reads "Albums" or "Composers" and its dropdown changes mode — so
it stays in one fixed spot while the content below swaps. The album grid becomes
grid-only; its shuffle and sort controls, and the composer sort, move into this
shared header.
